only reload if there are less characters than previously
This does improve search performance (at least on old phones), because before this commit every letter press was followed by 0.5-1 sec delay, and we don't need to reload the specifiers when only adding extra letters, because that only narrows the search results and doesn't widen them. This improved the smoothness of the search on my phone a lot.